fix issue with IO references in index
The problem was that interval_io[0] was sometime interpreted as
the reference to the declaration block, and sometime as the first
data sample.

This inconsistency is fixed by having interval_io[0] being the first
data block, and adding a declaration_block attribute which is the
reference to the stream's declaration block. Moreover, a read_one_block
now exists in order to have the reading share all the code, and more
importantly share all the safeguards.