fix reloading logfiles where the size of container changed
containers like vector<> can change size between machines. Fortunately, it
actually does not change the on-disk format marshalled by typelib.
Unfortunately, it *does* affect the type layouts (as for instance the size and
field offsets in structures that contain containers).

This commit makes use of the new Registry#resize method to adapt on the fly the
typelib registry, and avoid SIGSEGV'ing on such logfiles. It does so without
changing the file at all.