don't load the standard C++ types explicitely if it is not needed
Since september, generated TLBs contain all types. This fixes a few
portability issues between 32 and 64 bit machines, as std::string is
properly reported as having the size that it had in the file, instead
of the one of the host machine.

However, for backward compatibility reasons, the standard C++ types
*were* loaded in the registry. In practice, it meant that 32bit logs
containing std::string will make pocolog crash.

We now add the standard C++ types only for logs that are not self-contained
(i.e. "old logs").