bugfix: Set index to correct position if is requested position is after the stream
There was a problem, that if a position in the stream aligner was seeked
that was after the end of one of the aligned stream, the index of this
stream would remain at an undefined position. The obvious fix is to set
the index to the last position in the stream.