fixed indexing for multifile logs
Pocolog has the feature to combine multiple files into a single log stream
on the ruby side. The restructuring of the index mechanism missed to include
this feature. This lead to situations where file positions for samples where
provided for the wrong files.

This patch adds the support for multiple files per stream to the StreamIndex
class, and adds some additional checks on the way. This fix should be
transparent to the rest of the functionality, and the index files
automatically rebuild.