store time information in StreamIndex as numerical values
There is one Time object *per sample*. That means a lot of GC pressure
as the count of samples can easily reach a few millions. In addition,
the time is stored in a way that makes promotion to Bignum very unlikely
even on 32 bit systems.

This patch has two effects:
 - on all rubies we support, Time is a lot more costly than
   Integer to mark and sweep
 - marshalling to/from the indexes is a lot faster (we don't have
   to "unmarshal" the Time object anymore)
 - on ruby 2.1+, Time is a shady object (!). This patch therefore
   ensures that the index gets promoted to oldgen, thus reducing the
   GC pressure even more.

Not scientific, but running a log post-processing script that was taking
40s got reduced to 26s with this commit.