stream_aligner: normalize search behaviours in indexes, and the meaning of bounds in intervals
This commit moves the stream aligner in a [min, max[ meaning of intervals,
which is really really so much easier to handle. In particular, it
behaves much better when searching for a sample around a value, since
there is an easy past-the-end value (size) while there is none for
a last-sample-before behaviour (-1 ???)

It also tries to clear up some related unit tests.
unit test and document StreamAligner