avoid recursively sorting arrays in StreamAligner#add_stream
This changes the index-building method slightly so as to avoid sorting
arrays. The [time,stream_index] pair is basically encoded by having the
stream_index saved in the lower-significant part of a number where
the time is the higher-significant part. The position in stream does not
need to be encoded, as we can infer it when going over the sorted
"time codes".

This has a nice effect on performance.

The commit adds a fixture generator and a benchmark. The fixture
generator creates random-length, randomly-separated streams that
are then aligned by the benchmark.

With the following fixture:

Stream 0 [/int_t]
  68710 samples from Wed 20/04/2016 11:12:51 to Thu 21/04/2016 14:59:41 [27:46:49.887]
Stream 1 [/int_t]
  17733 samples from Wed 20/04/2016 11:12:54 to Thu 21/04/2016 14:59:35 [27:46:40.844]
Stream 2 [/int_t]
  70901 samples from Wed 20/04/2016 11:12:55 to Thu 21/04/2016 14:59:57 [27:47:02.172]
Stream 3 [/int_t]
  84717 samples from Wed 20/04/2016 11:12:58 to Thu 21/04/2016 15:02:54 [27:49:55.663]
Stream 4 [/int_t]
  95092 samples from Wed 20/04/2016 11:13:02 to Thu 21/04/2016 14:58:44 [27:45:41.658]
Stream 5 [/int_t]
  99013 samples from Wed 20/04/2016 11:13:07 to Thu 21/04/2016 15:00:33 [27:47:26.238]
Stream 6 [/int_t]
  66092 samples from Wed 20/04/2016 11:13:11 to Thu 21/04/2016 15:00:18 [27:47:06.478]
Stream 7 [/int_t]
  12899 samples from Wed 20/04/2016 11:13:14 to Thu 21/04/2016 14:55:24 [27:42:09.970]
Stream 8 [/int_t]
  37802 samples from Wed 20/04/2016 11:13:15 to Thu 21/04/2016 15:01:57 [27:48:42.509]
Stream 9 [/int_t]
  68587 samples from Wed 20/04/2016 11:13:17 to Thu 21/04/2016 14:59:55 [27:46:38.378]

The original code was doing

                           user     system      total        real
incremental           17.000000   0.060000  17.060000 ( 17.043634)
one-time               3.120000   0.000000   3.120000 (  3.114005)

The new one

                           user     system      total        real
incremental            3.110000   0.030000   3.140000 (  3.144666)
one-time               0.510000   0.000000   0.510000 (  0.513318)

For completeness, the code before the implementation of #add_streams

                           user     system      total        real
one-time               3.180000   0.040000   3.220000 (  3.215304)