upgrade: start the refactoring of rock-convert into rock-upgrade
The creation of a new tool is really there to avoid breaking existing
workflows. rock-export needed a rewrite badly (it could not handle
converter chains, was working exclusively on type names, had no good
way to add types, and was basically doing most of the hard work on the
hot path and had zero unit tests)

The general design here is to compute the conversion operations based
on the type model (i.e. once per stream) and then apply it on all values
in the stream. This commit implements the general mechanisms, it still
needs to be integrated into a CLI tool.