upgrade: implement a full-blown custom converter chain support
The code builds a graph of the possible conversions, and finds the
shortest path in it. If necessary, it tries to cast the source and
target type to something that has the same name *and* converters,
but the magic stops there. It won't for instance try to bridge a
gap between two declared converters.

From a CLI point of view, we'll need some debugging tools to find
these "holes" if they exist, and allow the user to add a deep cast
converter between the holes.