aligner: fix aligning empty streams
In general, there's no reason to not support this corner case.
In particular, Vizkit actually calls the aligner "as if" the
empty streams were aligned, but orocos/log skips the empty
streams. While this "magically" works now, there's not reason
it should - the refactored pocolog will raise if an unknown
stream is queried.