perf: flatten StreamIndex's internal data structure
It was an array of 3-tuples, which means we were paying a full object
(64 bytes + associated GC costs) for each single samples. Normalizing
decent logs (i.e. 24h or more) in syskit-log would halt to a crawl.

In addition to requiring a flatten to write indexes, and a processing
when loading.